pyproject.toml
[project]
name = "sensorpush_local"
version = "1.0.1"
description = "Local Bluetooth monitoring for SensorPush devices"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "homeassistant",
    "bleak-retry-connector>=3.4.0",
    # aiousbwatcher and pyserial are not used directly by this integration
    # but are hard-imported by homeassistant.components.usb (pulled in via
    # the bluetooth stack). HA's wheel does not declare them in install_requires
    # so pip will not install them transitively. Do not remove these.
    "aiousbwatcher",
    "pyserial",
]

[project.optional-dependencies]
test = [
    "pytest-homeassistant-custom-component",
    "pytest-asyncio",
    "pytest-cov",
    "argcomplete",
    "black",
    "flake8",
    "flake8-pyproject",
    "isort",
    "pylint",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = [".git"]
asyncio_mode = "auto"
log_cli = "true"
log_cli_level = "DEBUG"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:inotify is a Linux and FreeBSD API:UserWarning",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
